requirements.txt
django
djangorestframework
pytest
djangorestframework-simplejwt
